# pkg/openapi3/task_types.gen.go
// Package openapi3 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.8.3 DO NOT EDIT.
package openapi3

import (
	"time"
)

// Defines values for Priority.
const (
	PriorityHigh Priority = "high"

	PriorityLow Priority = "low"

	PriorityMedium Priority = "medium"

	PriorityNone Priority = "none"
)

// Dates defines model for Dates.
type Dates struct {
	Due   *time.Time `json:"due"`
	Start *time.Time `json:"start"`
}

// Priority defines model for Priority.
type Priority string

// Task defines model for Task.
type Task struct {
	Dates       *Dates    `json:"dates,omitempty"`
	Description *string   `json:"description,omitempty"`
	Id          *string   `json:"id,omitempty"`
	IsDone      *bool     `json:"is_done,omitempty"`
	Priority    *Priority `json:"priority,omitempty"`
}

// CreateTasksResponse defines model for CreateTasksResponse.
type CreateTasksResponse struct {
	Task *Task `json:"task,omitempty"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error *string `json:"error,omitempty"`
}

// ReadTasksResponse defines model for ReadTasksResponse.
type ReadTasksResponse struct {
	Task *Task `json:"task,omitempty"`
}

// SearchTasksResponse defines model for SearchTasksResponse.
type SearchTasksResponse struct {
	Tasks *[]Task `json:"tasks,omitempty"`
	Total *int64  `json:"total,omitempty"`
}

// CreateTasksRequest defines model for CreateTasksRequest.
type CreateTasksRequest struct {
	Dates       *Dates    `json:"dates,omitempty"`
	Description *string   `json:"description,omitempty"`
	Priority    *Priority `json:"priority,omitempty"`
}

// SearchTasksRequest defines model for SearchTasksRequest.
type SearchTasksRequest struct {
	Description *string   `json:"description"`
	From        *int64    `json:"from,omitempty"`
	IsDone      *bool     `json:"is_done"`
	Priority    *Priority `json:"priority,omitempty"`
	Size        *int64    `json:"size,omitempty"`
}

// UpdateTasksRequest defines model for UpdateTasksRequest.
type UpdateTasksRequest struct {
	Dates       *Dates    `json:"dates,omitempty"`
	Description *string   `json:"description,omitempty"`
	IsDone      *bool     `json:"is_done,omitempty"`
	Priority    *Priority `json:"priority,omitempty"`
}

// SearchTaskJSONRequestBody defines body for SearchTask for application/json ContentType.
type SearchTaskJSONRequestBody SearchTasksRequest

// UpdateTaskJSONRequestBody defines body for UpdateTask for application/json ContentType.
type UpdateTaskJSONRequestBody UpdateTasksRequest

// CreateTaskJSONRequestBody defines body for CreateTask for application/json ContentType.
type CreateTaskJSONRequestBody CreateTasksRequest